Upgrade changes, player instance property added, scriptables moved
- Upgrade class is now an IItem with proper constructor for Upgrade manager to work with.
- StatManager is now a property for Player Instance.
- Player Health setup delegate subscription with the maxHealth statTracker from stat manager.
- move definition scriptable objects moved to the resources folder.

+++ Fist Fishing/Assets/New Scripts/UpgradeManager.cs
@@ -40,7 +40,7 @@ public class UpgradeManager : MonoBehaviour
     /// </summary>
     void GenerateUpgrade()
     {
-
+        //Upgrade upgrade = new Upgrade()
     }
 
     /// <summary>
@@ -62,11 +62,11 @@ public class UpgradeManager : MonoBehaviour
         UpdateCosts?.Invoke(RecalculateCost);
     }
 
-    public delegate void CostChangeListener(System.Func<Dictionary<Stats, float>, float> updatefunction);
+    public delegate void CostChangeListener(System.Func<Dictionary<Stats, float>, int> updatefunction);
     public static CostChangeListener UpdateCosts;
 
-    static float RecalculateCost(Dictionary<Stats, float> statsModifier)
+    static int RecalculateCost(Dictionary<Stats, float> statsModifier)
     {
-        return 100.0f;
+        return 100;
     }
 }
